[MNY-286] SDK: Do not require connecting wallet in BuyWidget if receiverAddress is set (#8296)
<!--

## title your PR with this format: "[SDK/Dashboard/Portal] Feature/Fix: Concise title for the changes"

If you did not copy the branch name from Linear, paste the issue tag here (format is TEAM-0000):

## Notes for the reviewer

Anything important to call out? Be sure to also clarify these in your comments.

## How to test

Unit tests, playground, etc.

-->

<!-- start pr-codex -->

---

## PR-Codex overview

This PR focuses on modifying the `BuyWidget` component to allow users to bypass wallet connection if a `receiverAddress` is provided. It updates the logic for displaying connection prompts based on the presence of this address.

### Detailed summary

- Updated `BuyWidget` to not require wallet connection if `receiverAddress` is set.
- Removed `activeWalletInfo` dependency from the condition for displaying the `FundWallet`.
- Introduced `showConnectButton` variable to determine if the connect button should be shown.
- Simplified the rendering logic for the connect button in `FundWallet`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->